logica de obtencion de datos para el perfil

+++ src/main/java/com/focalizze/Focalizze/dto/ProfileResponseDto.java
@@ -10,7 +10,7 @@ public record ProfileResponseDto(
         Integer followers,
         Integer follow,
         Integer threadCount,
-        Integer threadsAvailableToday,
+        Long threadsAvailableToday,
         LocalDateTime registerDate
 
 ) {
